oasis: avoid linker message 'ld: -L../libs/quantor/ not found' and add -lpicosat
This commit intends to fix the warning

   ld: warning: directory not found for option '-L../libs/quantor/'

This warning is due to the fact that you must pass -L../libs/quantor/
in CCLib: field so that oasis can pass it to ocamlc and ocamlmklib. The
thing is that ocamlmklib actually remembers these options so that the
user of the stub library doesn't have to pass the -l and -L flags.

There is no (to my knowledge) way to avoid that and give specific
CFLAGS for build and for the CFLAGS that ocamlmklib will embed.

My workaround:
- In oasis, I pass the flag `-L. -lpicosat`
- I copy libpicosat.a and libquantor.a to _build to that the -L.
  will work at build time
- Even though -L. is embedded by ocamlmklib into qbf.cma (or wherever
  this information is stored), the -L. won't trigger a warning.