dune: remove -Lsrc/quantor
I did that in order to aboid the annoying ld warning that says that
src/quantor was not found when building an executable that uses
ocaml-qbf.

I also made sure there is no (bash ...) which was causing issues on
Windows + Cygwin: the command (bash %{make}) was picking up the wrong
make binary.