Pass correct path to linker with -L
Previously this code passed "-L." to the linker so that it would link
against the archives libpicosat.a and libquantor.a copied to src/quantor
during the build. Instead the linker was ignoring these files and using
the copies in the root directory of this project instead copied there by
the dune file in the root directory of this project. This is because the
linker is invoked from the workspace root (inside _build) rather than
the current directory.

Relevant issue: https://github.com/ocaml/dune/issues/7146

This happened to work by accident when building this project on its own
but in a setting where this project's source is vendored inside another
project the build would fail because the archives are placed into this
project's root directory but the linker is invoked from the workspace
root directory (ie. the outer project's root directory).

This patch works around the problem by using a rule to capture the
absolute path to the working directory which is then passed to the
linker.